sdk: derive config from chain id
