+++ providers/renegade-provider/config.ts
@@ -1,4 +1,9 @@
-import { createConfig, createStorage, getSDKConfig } from "@renegade-fi/react"
+import {
+  createConfig,
+  createStorage,
+  getSDKConfig,
+  isSupportedChainId,
+} from "@renegade-fi/react"
 
 import { env } from "@/env/client"
 import { cookieStorage } from "@/lib/cookie"
@@ -6,15 +11,18 @@ import { viemClient } from "@/lib/viem"
 
 export const sdkConfig = getSDKConfig(env.NEXT_PUBLIC_CHAIN_ID)
 
-export const config = createConfig({
-  chainId: sdkConfig.id,
-  darkPoolAddress: sdkConfig.darkpoolAddress,
-  priceReporterUrl: sdkConfig.priceReporterUrl,
-  relayerUrl: sdkConfig.relayerUrl,
-  ssr: true,
-  storage: createStorage({
-    storage: cookieStorage,
-  }),
-  // @sehyunc TODO: remove public client from config instantiation
-  viemClient: viemClient as any,
-})
+export const getConfigFromChainId = (chainId: number) => {
+  if (!isSupportedChainId(chainId)) return undefined
+  const sdkConfig = getSDKConfig(chainId)
+  return createConfig({
+    chainId,
+    darkPoolAddress: sdkConfig.darkpoolAddress,
+    priceReporterUrl: sdkConfig.priceReporterUrl,
+    relayerUrl: sdkConfig.relayerUrl,
+    ssr: true,
+    storage: createStorage({
+      storage: cookieStorage,
+    }),
+    viemClient: viemClient as any,
+  })
+}
